fix(v7): Move TotalComputationResources fct to where TransactionTrace is defined
